rating service

+++ src/main/java/dsw/projects/demo/movies/MovieDto.java
@@ -1,5 +1,5 @@
 package dsw.projects.demo.movies;
 
-public record MovieDto(Long id, String title, Float rating) {
+public record MovieDto(Long id, String title, Double rating) {
 
 }
